+++ Assets/Scripts/Expression/Event/ICommonEventCommandsRepository.cs
@@ -2,14 +2,14 @@ using System;
 using Util.Wolf;
 using UnityEngine;
 
-namespace Expression.Map.MapEvent
+namespace Expression.Event
 {
     /// <summary>
     /// コモンイベントのコマンド群を呼び出すためのインターフェース
     /// </summary>
     public interface ICommonEventCommandsRepository
     {
-        public EventCommandBase[] GetCommands(int commonEventId);
+        public CommonEvent GetEvent(int commonEventId);
 
         public int GetCount();
     }
